+++ scm/config/security/__init__.py
@@ -1,0 +1,3 @@
+# scm/config/security/__init__.py
+
+from .anti_spyware_profiles import AntiSpywareProfile
