+++ scm/models/__init__.py
@@ -1,10 +1,3 @@
 # scm/models/__init__.py
 
-from .address import AddressRequestModel, AddressResponseModel
-from .address_group import AddressGroupRequestModel, AddressGroupResponseModel
-from .application import ApplicationRequestModel, ApplicationResponseModel
-from .application_group import (
-    ApplicationGroupRequestModel,
-    ApplicationGroupResponseModel,
-)
-from .service import ServiceRequestModel, ServiceResponseModel
+from .auth import AuthRequestModel
