+++ scm/models/objects/__init__.py
@@ -1,0 +1,13 @@
+# scm/models/objects/__init__.py
+
+from .address import AddressRequestModel, AddressResponseModel
+from .address_group import AddressGroupRequestModel, AddressGroupResponseModel
+from .application import (
+    ApplicationRequestModel,
+    ApplicationResponseModel,
+)
+from .application_group import (
+    ApplicationGroupRequestModel,
+    ApplicationGroupResponseModel,
+)
+from .service import ServiceRequestModel, ServiceResponseModel
